Be208 authentication

+++ backend/src/routes/users.js
@@ -19,18 +19,24 @@ router.get('/:userId', async function (req, res, next) {
   res.send(user)
 })
 
-// create a new user
+// create a user with a name
 router.post('/', async function (req, res, next) {
-  const { name, age } = req.body
-  const newUser = await User.create({ name, age })
+  const { name, age, email, password } = req.body
 
-  res.send(newUser)
+  const user = await User.register({ name, email, age }, password)
+
+  res.send(user)
 })
 
 // update a user
 router.put('/:userId', async function (req, res, next) {
   const { userId } = req.params
   const { newValues } = req.body
+
+  if (!req.user) return next({ status: 404, message: 'User not found' })
+
+  if (req.user._id.toString() !== userId)
+    return next({ status: 403, message: 'You are not allowed to update this user' })
 
   const updatedUser = await User.findByIdAndUpdate(userId, { $set: newValues }, { new: true })
 
@@ -40,6 +46,11 @@ router.put('/:userId', async function (req, res, next) {
 // delete a user
 router.delete('/:userId', async function (req, res, next) {
   const { userId } = req.params
+
+  if (!req.user) return next({ status: 404, message: 'User not found' })
+
+  if (req.user._id.toString() !== userId)
+    return next({ status: 403, message: 'You are not allowed to delete this user' })
 
   await User.findByIdAndDelete(userId)
